packages/python/neuronpedia-webapp-client/pyproject.toml
[tool.poetry]
name = "neuronpedia"
version = "1.0.22"
description = "Neuronpedia"
authors = ["johnny <johnny@neuronpedia.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120